requirements.txt
camtools
einops
lpips
tensorboard
tqdm
transformers
omegaconf
open-clip-torch
open3d
opencv-python
Pillow
plyfile
pytorch-lightning
PyYAML
ipykernel
trimesh
roma
./submodules/diff-gaussian-rasterization
./submodules/diff-gaussian-rasterization-w-pose
./submodules/simple-knn
./submodules/pytorch3d
./submodules/minLoRA
./submodules/CLIP
./submodules/segment-anything
